requirements.txt
autopep8==1.7.0
flake8==5.0.4
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyserial==3.5
toml==0.10.2
